backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.9.1
google-cloud-vision==3.5.0
google-generativeai==0.7.2
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.0
httpx==0.27.0
pytest==7.4.0
